requirements.txt
altgraph==0.17
attrs==21.2.0
auto-py-to-exe==2.10.0
autobahn==21.3.1
Automat==20.2.0
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.14.6
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
hkdf==0.0.3
humanize==3.10.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
magic-wormhole==0.12.0
magic-wormhole-mailbox-server==0.4.1
magic-wormhole-transit-relay==0.2.1
mock==4.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
qt5reactor==0.6.3
service-identity==21.1.0
six==1.16.0
spake2==0.8
tqdm==4.62.0
Twisted==21.7.0
txaio==21.2.1
txtorcon==20.0.0
typing-extensions==3.10.0.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
